Fix integration tests
The integration tests were failing due to multiple issues.

Multiple tests were using the same container name (`testing`),
which meant some of the tests weren't running. Now each container
has a unique name.

The same port (9000) was being used for multiple tests. This
wasn't showing as an issue as these tests also used the same
container name. The ports are now unique per test, with them
being number sequentially to avoid any conflicts.

Only some of the running containers were terminated when the suite
finished executing. This caused many tests to not run properly on
subsequent invocation of the suite. Now all containers are
terminated when the tests complete.

The timeout tests weren't waiting for the container be ready. This was
caused by the order of operations being wrong, with the sleep step
occurring after the request to the endpoint, not before.

This commit fixes #47.