test: Add delay for time-related arm64 tests (#138)
The tests on GitHub run on x86 instances (because arm64 instances don't
have Docker installed) so when running the arm64 tests, invokes take
longer because of the cross-architecture emulation. This caused that
some tests that check remaining time in the function were not landing
on the correct time range.

It's unknown why this delay started manifesting more consistently, we might
want to find a better solution in the future.